Refactoring in preparation to the target report API.
