+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/CacheConfiguration.java
@@ -1,7 +1,6 @@
 package org.opentestsystem.rdw.olap;
 
 import org.opentestsystem.rdw.common.model.AssessmentType;
-import org.opentestsystem.rdw.olap.model.Target;
 import org.opentestsystem.rdw.olap.repository.ActiveAssessmentRepository;
 import org.opentestsystem.rdw.olap.repository.AdministrationConditionRepository;
 import org.opentestsystem.rdw.olap.repository.AssessmentGradeRepository;
@@ -26,9 +25,9 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.annotation.Scheduled;
 
 import static org.opentestsystem.rdw.olap.model.AggregateServiceCache.AdministrationConditions;
-import static org.opentestsystem.rdw.olap.model.AggregateServiceCache.Assessments;
 import static org.opentestsystem.rdw.olap.model.AggregateServiceCache.AssessmentGrades;
 import static org.opentestsystem.rdw.olap.model.AggregateServiceCache.AssessmentTypes;
+import static org.opentestsystem.rdw.olap.model.AggregateServiceCache.Assessments;
 import static org.opentestsystem.rdw.olap.model.AggregateServiceCache.Booleans;
 import static org.opentestsystem.rdw.olap.model.AggregateServiceCache.Claims;
 import static org.opentestsystem.rdw.olap.model.AggregateServiceCache.CompletenessCodes;
@@ -64,10 +63,10 @@ public class CacheConfiguration {
 
     @Autowired
     private CompletenessRepository completenessRepository;
-    
+
     @Autowired
     private ClaimRepository claimRepository;
-    
+
     @Autowired
     private ElasRepository elasRepository;
