+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/model/ReportTemplate.java
@@ -5,6 +5,7 @@ import org.opentestsystem.rdw.reporting.common.model.Dimension;
 import org.opentestsystem.rdw.reporting.common.model.District;
 import org.opentestsystem.rdw.reporting.common.model.Organization;
 import org.opentestsystem.rdw.reporting.common.model.School;
+import org.opentestsystem.rdw.reporting.common.model.Target;
 
 import java.util.Collection;
 import java.util.Map;
