+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/repository/TargetRepository.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.rdw.olap.repository;
 
 import org.opentestsystem.rdw.olap.model.Claim;
-import org.opentestsystem.rdw.olap.model.Target;
+import org.opentestsystem.rdw.reporting.common.model.Target;
 
 import java.util.Map;
 import java.util.Set;
