+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/repository/impl/JdbcTargetReportRepository.java
@@ -3,7 +3,7 @@ package org.opentestsystem.rdw.olap.repository.impl;
 
 import org.opentestsystem.rdw.olap.model.ReportRowIdentity;
 import org.opentestsystem.rdw.olap.model.ReportTemplate;
-import org.opentestsystem.rdw.olap.model.Target;
+import org.opentestsystem.rdw.reporting.common.model.Target;
 import org.opentestsystem.rdw.olap.model.TargetOrganizationQuery;
 import org.opentestsystem.rdw.olap.model.TargetReportResult;
 import org.opentestsystem.rdw.olap.model.TargetRowIdentity;
