+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/repository/impl/JdbcTargetRepository.java
@@ -1,15 +1,14 @@
 package org.opentestsystem.rdw.olap.repository.impl;
 
 import org.opentestsystem.rdw.olap.model.AggregateServiceCache;
-import org.opentestsystem.rdw.olap.model.Target;
 import org.opentestsystem.rdw.olap.repository.TargetRepository;
+import org.opentestsystem.rdw.reporting.common.jdbc.Targets;
+import org.opentestsystem.rdw.reporting.common.model.Target;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.Map;
 import java.util.Set;
 
@@ -30,16 +29,6 @@ class JdbcTargetRepository implements TargetRepository {
     @Cacheable(AggregateServiceCache.Targets)
     @Override
     public Map<Integer, Set<Target>> findAll() {
-        return template.query(findAllQuery, (row, index) -> map(row)).stream().collect(groupingBy(Target::getAssessmentId, toSet()));
-    }
-
-    private static Target map(final ResultSet row) throws SQLException {
-        return Target.builder()
-                .id(row.getInt("id"))
-                .assessmentId(row.getInt("asmt_id"))
-                .naturalId(row.getString("natural_id"))
-                .claimCode(row.getString("claim_code"))
-                .includeInReport(row.getBoolean("include_in_report"))
-                .build();
+        return template.query(findAllQuery, (row, index) -> Targets.map(row)).stream().collect(groupingBy(Target::getAssessmentId, toSet()));
     }
 }
