+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/sqlbuilder/impl/QueryProviderRepositoryHelper.java
@@ -2,7 +2,7 @@ package org.opentestsystem.rdw.olap.sqlbuilder.impl;
 
 import com.google.common.collect.ImmutableList;
 import org.opentestsystem.rdw.olap.model.OrganizationTypeQuery;
-import org.opentestsystem.rdw.olap.model.Target;
+import org.opentestsystem.rdw.reporting.common.model.Target;
 import org.opentestsystem.rdw.olap.repository.ActiveAssessmentRepository;
 import org.opentestsystem.rdw.olap.repository.AdministrationConditionRepository;
 import org.opentestsystem.rdw.olap.repository.AssessmentGradeRepository;
