+++ aggregate-service/src/test/java/org/opentestsystem/rdw/olap/repository/impl/JdbcTargetRepositoryRST.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.rdw.olap.repository.impl;
 
 import org.junit.Test;
-import org.opentestsystem.rdw.olap.model.Target;
+import org.opentestsystem.rdw.reporting.common.model.Target;
 import org.opentestsystem.rdw.olap.repository.TargetRepository;
 import org.opentestsystem.rdw.reporting.common.test.CachingTest;
 import org.springframework.beans.factory.annotation.Autowired;
