+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/jdbc/Targets.java
@@ -1,0 +1,22 @@
+package org.opentestsystem.rdw.reporting.common.jdbc;
+
+import org.opentestsystem.rdw.reporting.common.model.Target;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+/**
+ * Holds utility methods for mapping SQL query results to {@link Target}
+ */
+public final class Targets {
+
+    public static Target map(final ResultSet row) throws SQLException {
+        return Target.builder()
+                .id(row.getInt("id"))
+                .assessmentId(row.getInt("asmt_id"))
+                .naturalId(row.getString("natural_id"))
+                .claimCode(row.getString("claim_code"))
+                .includeInReport(row.getBoolean("include_in_report"))
+                .build();
+    }
+}
