+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/Target.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.olap.model;
+package org.opentestsystem.rdw.reporting.common.model;
 
 /**
  * Represents a target
