+++ report-processor/src/main/java/org/opentestsystem/rdw/reporting/processor/repository/JdbcIcaSummativeReportRepository.java
@@ -4,6 +4,7 @@ import org.opentestsystem.rdw.common.model.AssessmentType;
 import org.opentestsystem.rdw.common.model.Subject;
 import org.opentestsystem.rdw.reporting.common.configuration.ReportingSystemSettings;
 import org.opentestsystem.rdw.reporting.common.jdbc.Assessments;
+import org.opentestsystem.rdw.reporting.common.jdbc.ClaimScores;
 import org.opentestsystem.rdw.reporting.common.jdbc.Exams;
 import org.opentestsystem.rdw.reporting.common.jdbc.SecurityParameterProvider;
 import org.opentestsystem.rdw.reporting.common.jdbc.Students;
@@ -43,8 +44,8 @@ class JdbcIcaSummativeReportRepository implements IcaSummativeReportRepository {
 
     @Autowired
     JdbcIcaSummativeReportRepository(final NamedParameterJdbcTemplate template,
-                            final SecurityParameterProvider securityParameterProvider,
-                            final ReportingSystemSettings systemSettings) {
+                                     final SecurityParameterProvider securityParameterProvider,
+                                     final ReportingSystemSettings systemSettings) {
         this.template = template;
         this.securityParameterProvider = securityParameterProvider;
         this.systemSettings = systemSettings;
@@ -69,7 +70,7 @@ class JdbcIcaSummativeReportRepository implements IcaSummativeReportRepository {
                 (row) -> {
                     final IcaSummativeReport report = IcaSummativeReport.builder()
                             .assessment(Assessments.mapWithCutPoints(row, Assessment.builder(), "asmt_").build())
-                            .exam(Exams.map(row, Exam.builder()).build())
+                            .exam(Exams.map(row, Exam.builder()).claimScaleScores(ClaimScores.map(row)).build())
                             .student(Students.map(row, Student.builder(), "student_").build())
                             .studentEnrollment(StudentEnrollments.map(row, systemSettings.getState().getCode()))
                             .build();
