+++ report-processor/src/test/java/org/opentestsystem/rdw/reporting/processor/repository/JdbcStudentExamPermissionRepository.java
@@ -2,6 +2,7 @@ package org.opentestsystem.rdw.reporting.processor.repository;
 
 import com.google.common.collect.ImmutableMap;
 import org.opentestsystem.rdw.reporting.common.jdbc.Assessments;
+import org.opentestsystem.rdw.reporting.common.jdbc.ClaimScores;
 import org.opentestsystem.rdw.reporting.common.jdbc.Exams;
 import org.opentestsystem.rdw.reporting.common.jdbc.SecurityParameterProvider;
 import org.opentestsystem.rdw.reporting.common.model.Assessment;
@@ -43,7 +44,9 @@ class JdbcStudentExamPermissionRepository implements StudentExamPermissionReposi
                 (row, index) -> {
                     //Build the wrapper
                     return StudentHistoryExamWrapper.builder()
-                            .exam(Exams.map(row, Exam.builder(), permissionSource.getPermissionsById()).build())
+                            .exam(Exams.map(row, Exam.builder(), permissionSource.getPermissionsById())
+                                    .claimScaleScores(ClaimScores.map(row))
+                                    .build())
                             .assessment(Assessments.map(row, Assessment.builder(), "asmt_").build())
                             .build();
                 });
