+++ reporting-service/src/main/java/org/opentestsystem/rdw/reporting/exam/AbstractExamRepository.java
@@ -1,5 +1,6 @@
 package org.opentestsystem.rdw.reporting.exam;
 
+import org.opentestsystem.rdw.reporting.common.jdbc.ClaimScores;
 import org.opentestsystem.rdw.reporting.common.jdbc.Exams;
 import org.opentestsystem.rdw.reporting.common.jdbc.SecurityParameterProvider;
 import org.opentestsystem.rdw.reporting.common.jdbc.StudentContexts;
@@ -65,7 +66,8 @@ public abstract class AbstractExamRepository<T extends AbstractExamSearch> imple
             final Exam.Builder examBuilder = Exam.builder()
                     .studentContext(StudentContexts.map(resultSet, StudentContext.builder()).build());
 
-            examBuilders.add(Exams.map(resultSet, examBuilder, permissionSource.getPermissionsById()));
+            examBuilders.add(Exams.map(resultSet, examBuilder, permissionSource.getPermissionsById())
+                    .claimScaleScores(ClaimScores.map(resultSet)));
             studentIds.add(resultSet.getLong("student_id"));
         }
